test: 사전 주문 결제 통합테스트

+++ service-order/src/main/kotlin/me/bread/order/application/external/PaymentApi.kt
@@ -1,8 +1,7 @@
 package me.bread.order.application.external
 
 import me.bread.order.application.model.TossPayment
-import java.math.BigDecimal
 
 interface PaymentApi {
-    fun execute(orderId: Long, paymentId: String, amount: BigDecimal): TossPayment
+    fun execute(orderId: Long, paymentId: String, amount: Long): TossPayment
 }
