+++ service-order/src/main/kotlin/me/bread/order/application/model/TossPayment.kt
@@ -1,8 +1,6 @@
 package me.bread.order.application.model
 
-import java.math.BigDecimal
-
 data class TossPayment(
     val orderId: Long,
-    val amount: BigDecimal,
+    val amount: Long,
 )
