+++ service-order/src/main/kotlin/me/bread/order/presentation/rest/request/PaymentRequest.kt
@@ -1,5 +1,3 @@
 package me.bread.order.presentation.rest.request
 
-import java.math.BigDecimal
-
-data class PaymentRequest(val orderId: Long, val paymentKey: String, val amount: BigDecimal)
+data class PaymentRequest(val orderId: Long, val paymentKey: String, val amount: Long)
